+++ src/main/java/school/faang/user_service/controller/RecommendationController.java
@@ -3,7 +3,7 @@ package school.faang.user_service.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import school.faang.user_service.exception.DataValidationException;
-import school.faang.user_service.repository.recommendation.RecommendationDto;
+import school.faang.user_service.dto.recommendation.RecommendationDto;
 import school.faang.user_service.service.RecommendationService;
 
 import java.util.List;
