+++ src/main/java/school/faang/user_service/mapper/RecommendationMapper.java
@@ -4,7 +4,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
 import school.faang.user_service.entity.recommendation.Recommendation;
-import school.faang.user_service.repository.recommendation.RecommendationDto;
+import school.faang.user_service.dto.recommendation.RecommendationDto;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface RecommendationMapper {
